+++ magik-typed-checks/src/test/java/nl/ramsolutions/sw/magik/typedchecks/checks/SwChar16VectorEvaluateInvocationTypedCheckTest.java
@@ -9,13 +9,13 @@ import nl.ramsolutions.sw.magik.checks.MagikIssue;
 import nl.ramsolutions.sw.magik.typedchecks.MagikTypedCheck;
 import org.junit.jupiter.api.Test;
 
-class SwChar16VectorEvaluateInvocationCheckTest extends MagikTypedCheckTestBase {
+class SwChar16VectorEvaluateInvocationTypedCheckTest extends MagikTypedCheckTestBase {
 
   @Test
   void testUseOfSwChar16VectorEvaluate() {
     final String code = "'abc'.evaluate()";
     final IDefinitionKeeper definitionKeeper = new DefinitionKeeper();
-    final MagikTypedCheck check = new SwChar16VectorEvaluateInvocationCheck();
+    final MagikTypedCheck check = new SwChar16VectorEvaluateInvocationTypedCheck();
     final List<MagikIssue> checkResults = this.runCheck(code, definitionKeeper, check);
     assertThat(checkResults).hasSize(1);
   }
@@ -24,7 +24,7 @@ class SwChar16VectorEvaluateInvocationCheckTest extends MagikTypedCheckTestBase 
   void testUseOfOtherEvaluate() {
     final String code = "100.evaluate()";
     final IDefinitionKeeper definitionKeeper = new DefinitionKeeper();
-    final MagikTypedCheck check = new SwChar16VectorEvaluateInvocationCheck();
+    final MagikTypedCheck check = new SwChar16VectorEvaluateInvocationTypedCheck();
     final List<MagikIssue> checkResults = this.runCheck(code, definitionKeeper, check);
     assertThat(checkResults).isEmpty();
   }