+++ CHANGES.md
@@ -9,6 +9,7 @@
 - Refactor `MagikFileScanner` & friends to be shared.
 - Add `magik-typed-lint` to check magik typing from CLI/during CI.
 - Fix not indexing `module.def` files.
+- Rename SwChar16VectorEvaluateInvocationCheck to SwChar16VectorEvaluateInvocationTypedCheck.
 - Several fixes.
 
 0.10.1 (2024-08-14)